feat: v2 sovereign agent economy — CLAUDE.md constitution, landing scenes, wallet package, contract interfaces
Rewrite CLAUDE.md as project constitution establishing the Sovereign Agent Economy
identity, Lead Protocol Architect role, and x402 economic design principles.

- Add 15+ cinematic landing page scene components (data vortex, reactor, terminal, etc.)
- Add packages/wallet for MPC wallet management (crypto, store, AgentKit integration)
- Add contract ABIs for all 11 deployed contracts (AgentFactory through X402PaymentGate)
- Add agent deploy flow (deploy-button, use-agent-deploy hook, activate API route)
- Update smart contracts: FeeSplitter v2, ISwapRouter, IVirtualsFactory interfaces
- Update Prisma schema with financial tracking models
- Fix next.config.ts transpilePackages for viem/AgentKit compatibility

+++ apps/web/hooks/finance/use-staking-rewards.ts
@@ -263,7 +263,7 @@ export function usePoolAPY(pid: number) {
 
   // Annual RUN for this pool = (runPerSecond * allocPoint / totalAllocPoint) * seconds_per_year
   const SECONDS_PER_YEAR = 365n * 24n * 60n * 60n;
-  const poolRunPerYear = (runPerSecond * pool.allocPoint * SECONDS_PER_YEAR) / totalAllocPoint;
+  const poolRunPerYear = ((runPerSecond as bigint) * pool.allocPoint * SECONDS_PER_YEAR) / (totalAllocPoint as bigint);
 
   // APY = (annual_rewards / total_staked) * 100
   // Using 1e18 precision since both are in wei
